Extract AbstractCodecConfigurer base class
Issue: SPR-15247
